+++ src/toolkitgroupview.c
@@ -398,8 +398,7 @@ toolkitgroupview_build_toolitem(void *a, void *b, void *c)
 	Toolkitgroupview *kitgview = TOOLKITGROUPVIEW(c);
 
 	if (toolitem->tool &&
-		(kitgview->show_all || MODEL(toolitem->tool)->display) &&
-		(kitgview->show_all || !toolitem->is_separator))
+		(kitgview->show_all || MODEL(toolitem->tool)->display))
 		toolkitgroupview_add_toolitem(kitgview, store, toolitem);
 
 	return NULL;
@@ -555,8 +554,6 @@ toolkitgroupview_bind_browse_item(GtkListItemFactory *factory,
 		gtk_widget_set_visible(pin, TRUE);
 		gtk_label_set_xalign(GTK_LABEL(label), 0.5);
 		g_object_set_qdata(G_OBJECT(button), node_quark, parent);
-
-		// note the pin widget on the page
 	}
 	else {
 		gtk_label_set_xalign(GTK_LABEL(label), 0.0);
@@ -572,10 +569,16 @@ toolkitgroupview_bind_browse_item(GtkListItemFactory *factory,
 	}
 
 	if (node->toolitem &&
-		node->toolitem->is_separator)
+		node->toolitem->is_separator) {
 		gtk_widget_set_sensitive(button, FALSE);
+		gtk_widget_add_css_class(button, "toolkitgroupview-sep");
+	}
+	else
+		gtk_widget_remove_css_class(button, "toolkitgroupview-sep");
 
-	gtk_widget_remove_css_class(gtk_widget_get_parent(button), "activatable");
+	GtkWidget *enclosing_box = gtk_widget_get_parent(button);
+	GtkWidget *enclosing_item = gtk_widget_get_parent(enclosing_box);
+	gtk_widget_remove_css_class(enclosing_item, "activatable");
 }
 
 static GtkWidget *