+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - improve Text
 - improve scroll to row on error
+- improve styling of toolkitgroupview
+- add Menuseparator support
 
 ## 9.0.12 2025/08/21